+++ examples/movies/src/app.tsx
@@ -3,14 +3,7 @@ import { useRequest } from 'react-rest-request';
 import { MoviesEndpoint, MoviesResponse } from './endpoint';
 
 export default function App() {
-    const [movies, { data, loading }] = useRequest<MoviesResponse>(MoviesEndpoint);
-
-    React.useEffect(
-        () => {
-            movies();
-        },
-        [movies]
-    );
+    const { data, loading } = useRequest<MoviesResponse>(MoviesEndpoint);
 
     return !data ? (
         <div>{ loading ? 'Loading...' : 'Something went wrong' }</div>
