+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from './endpoint';
 export * from './client';
 export * from './client-hook';
+export * from './lazy-request-hook';
 export * from './request-hook';
 export * from './request-context';
 export * from './reducer';
